♻️ Refactor: Extract UserSearch to new component
