+++ src/components/UserSearch/UserSearch.scss
@@ -1,0 +1,67 @@
+@import "../../App";
+
+.search-form {
+    position: relative;
+}
+
+.search-dropdown {
+    position: absolute;
+    top: calc(100% + 1rem);
+    left: -2rem;
+    right: -2rem;
+    z-index: 10;
+    outline: 2px solid $color-dark;
+    padding: 0.5rem;
+    background-color: $color-xdark;
+    border-radius: 1rem;
+}
+
+.search-result {
+    position: relative;
+    display: flex;
+    flex-direction: row;
+    gap: 1rem;
+    padding: 0.5rem;
+    border-radius: 0.8rem;
+
+    &:not(:last-child) {
+        &::after {
+            content: '';
+            position: absolute;
+            bottom: 0;
+            width: 12rem;
+            height: 0;
+            left: 0;
+            right: 0;
+            margin: 0 auto;
+            border-top: 1px solid $color-dark;
+        }
+    }
+
+    &:hover {
+        background-color: $color-dark;
+    }
+
+    >div {
+        display: flex;
+        flex-direction: column;
+        gap: 0.25rem;
+    }
+
+    img {
+        width: 3rem;
+        height: 3rem;
+        border-radius: 50%;
+    }
+
+    h4,
+    p {
+        padding: 0;
+        margin: 0;
+    }
+
+    .username {
+        opacity: 0.75;
+    }
+
+}
